kes  Fix a couple of compiler warnings in wx-console kes  Apply the immutable patch for FreeBSD hardlinked files supplied      by Olivier Lehmann.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@3822 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/wx-console/wxbconfigpanel.cpp
@@ -137,13 +137,13 @@ void wxbConfigParam::SetValue(wxString str) {
    case choice:
       if (choicectrl) {
          int k;
-         for (k = 0; k < choicectrl->GetCount(); k++) {
+         for (k = 0; k < (int)choicectrl->GetCount(); k++) {
             if (str == choicectrl->GetString(k)) {
                choicectrl->SetSelection(k);
                break;
             }
          }
-         if (k == choicectrl->GetCount()) { // Should never happen
+         if (k == (int)choicectrl->GetCount()) { // Should never happen
             choicectrl->Append(str);
             choicectrl->SetSelection(k);
          }
